refactor(onboard): keep reasoning helpers out of entrypoint
